refactor: better database field names and fix delete on set
