+++ src/commands/moderation/sticky.ts
@@ -5,7 +5,7 @@ import { Roles, Channels, CommandCategory } from '../../constants';
 import { makeEmbed } from '../../lib/embed';
 import { getConn } from '../../lib/db';
 import StickyMessage from '../../lib/schemas/stickyMessageSchema';
-import { stickyMessageEmbed } from '../../lib/stickyEmbed';
+import { stickyMessageEmbed } from '../../lib/stickyMessageEmbed';
 
 const permittedRoles = [
     Roles.ADMIN_TEAM,
@@ -97,7 +97,7 @@ const notFoundEmbed = (action: string, channel: string) => makeEmbed({
     color: Colors.Red,
 });
 
-const stickyMessageEmbedField = (updatedDate: string, moderator: string, message: string, timeInterval: string, messageCount: string, channel: string): EmbedField[] => [
+const stickyMessageEmbedField = (updatedTimestamp: string, moderator: string, message: string, timeInterval: string, messageCount: string, channel: string): EmbedField[] => [
     {
         inline: true,
         name: 'Channel',
@@ -110,8 +110,8 @@ const stickyMessageEmbedField = (updatedDate: string, moderator: string, message
     },
     {
         inline: true,
-        name: 'Last updated date',
-        value: updatedDate,
+        name: 'Last updated timestamp',
+        value: updatedTimestamp,
     },
     {
         inline: true,
@@ -181,7 +181,14 @@ export const sticky: CommandDefinition = {
             }
             stickyMessage.message = message;
             stickyMessage.moderator = author.toString();
-            stickyMessage.updatedDate = new Date();
+            stickyMessage.updatedTimestamp = new Date();
+
+            const previousSticky = stickyMessage.lastPostedId ? await msg.channel.messages.fetch(stickyMessage.lastPostedId) : null;
+            if (previousSticky) {
+                previousSticky.delete();
+            }
+            const currentSticky = await msg.channel.send({ embeds: [stickyMessageEmbed(stickyMessage.message)] });
+            stickyMessage.lastPostedId = currentSticky.id;
 
             try {
                 stickyMessage.save();
@@ -190,12 +197,11 @@ export const sticky: CommandDefinition = {
             }
 
             try {
-                modLogsChannel.send({ embeds: [channelEmbed('Set', stickyMessageEmbedField(moment(stickyMessage.updatedDate).utcOffset(0).format('YYYY-MM-DD HH:mm:ss'), stickyMessage.moderator, stickyMessage.message, `${stickyMessage.timeInterval}`, `${stickyMessage.messageCount}`, stickyMessage.channelId), Colors.Green)] });
+                modLogsChannel.send({ embeds: [channelEmbed('Set', stickyMessageEmbedField(moment(stickyMessage.updatedTimestamp).utcOffset(0).format('YYYY-MM-DD HH:mm:ss'), stickyMessage.moderator, stickyMessage.message, `${stickyMessage.timeInterval}`, `${stickyMessage.messageCount}`, stickyMessage.channelId), Colors.Green)] });
             } catch {
                 msg.channel.send({ embeds: [noChannelEmbed('Set', 'Mod Log')] });
             }
 
-            msg.channel.send({ embeds: [stickyMessageEmbed(stickyMessage.message)] });
             return msg.react('✅');
         }
 
@@ -216,7 +222,7 @@ export const sticky: CommandDefinition = {
             }
             stickyMessage.messageCount = parseInt(messageCount);
             stickyMessage.moderator = author.toString();
-            stickyMessage.updatedDate = new Date();
+            stickyMessage.updatedTimestamp = new Date();
 
             try {
                 stickyMessage.save();
@@ -225,7 +231,7 @@ export const sticky: CommandDefinition = {
             }
 
             try {
-                modLogsChannel.send({ embeds: [channelEmbed('Configure minimum message count', stickyMessageEmbedField(moment(stickyMessage.updatedDate).utcOffset(0).format('YYYY-MM-DD HH:mm:ss'), stickyMessage.moderator, stickyMessage.message, `${stickyMessage.timeInterval}`, `${stickyMessage.messageCount}`, stickyMessage.channelId), Colors.Green)] });
+                modLogsChannel.send({ embeds: [channelEmbed('Configure minimum message count', stickyMessageEmbedField(moment(stickyMessage.updatedTimestamp).utcOffset(0).format('YYYY-MM-DD HH:mm:ss'), stickyMessage.moderator, stickyMessage.message, `${stickyMessage.timeInterval}`, `${stickyMessage.messageCount}`, stickyMessage.channelId), Colors.Green)] });
             } catch {
                 msg.channel.send({ embeds: [noChannelEmbed('Configure minimum message count', 'Mod Log')] });
             }
@@ -250,7 +256,7 @@ export const sticky: CommandDefinition = {
             }
             stickyMessage.timeInterval = parseInt(timeInterval);
             stickyMessage.moderator = author.toString();
-            stickyMessage.updatedDate = new Date();
+            stickyMessage.updatedTimestamp = new Date();
 
             try {
                 stickyMessage.save();
@@ -259,7 +265,7 @@ export const sticky: CommandDefinition = {
             }
 
             try {
-                modLogsChannel.send({ embeds: [channelEmbed('Configure minimum message count', stickyMessageEmbedField(moment(stickyMessage.updatedDate).utcOffset(0).format('YYYY-MM-DD HH:mm:ss'), stickyMessage.moderator, stickyMessage.message, `${stickyMessage.timeInterval}`, `${stickyMessage.messageCount}`, stickyMessage.channelId), Colors.Green)] });
+                modLogsChannel.send({ embeds: [channelEmbed('Configure minimum message count', stickyMessageEmbedField(moment(stickyMessage.updatedTimestamp).utcOffset(0).format('YYYY-MM-DD HH:mm:ss'), stickyMessage.moderator, stickyMessage.message, `${stickyMessage.timeInterval}`, `${stickyMessage.messageCount}`, stickyMessage.channelId), Colors.Green)] });
             } catch {
                 msg.channel.send({ embeds: [noChannelEmbed('Configure minimum message count', 'Mod Log')] });
             }
@@ -282,7 +288,7 @@ export const sticky: CommandDefinition = {
             }
 
             try {
-                modLogsChannel.send({ embeds: [channelEmbed('Delete', stickyMessageEmbedField(moment(stickyMessage.updatedDate).utcOffset(0).format('YYYY-MM-DD HH:mm:ss'), stickyMessage.moderator, stickyMessage.message, `${stickyMessage.timeInterval}`, `${stickyMessage.messageCount}`, stickyMessage.channelId), Colors.Red)] });
+                modLogsChannel.send({ embeds: [channelEmbed('Delete', stickyMessageEmbedField(moment(stickyMessage.updatedTimestamp).utcOffset(0).format('YYYY-MM-DD HH:mm:ss'), stickyMessage.moderator, stickyMessage.message, `${stickyMessage.timeInterval}`, `${stickyMessage.messageCount}`, stickyMessage.channelId), Colors.Red)] });
             } catch {
                 msg.channel.send({ embeds: [noChannelEmbed('Delete', 'Mod Log')] });
             }
@@ -304,7 +310,7 @@ export const sticky: CommandDefinition = {
             }
 
             const [stickyMessage] = searchResult;
-            return msg.channel.send({ embeds: [infoEmbed(stickyMessageEmbedField(moment(stickyMessage.updatedDate).utcOffset(0).format('YYYY-MM-DD HH:mm:ss'), stickyMessage.moderator, stickyMessage.message, `${stickyMessage.timeInterval}`, `${stickyMessage.messageCount}`, stickyMessage.channelId))] });
+            return msg.channel.send({ embeds: [infoEmbed(stickyMessageEmbedField(moment(stickyMessage.updatedTimestamp).utcOffset(0).format('YYYY-MM-DD HH:mm:ss'), stickyMessage.moderator, stickyMessage.message, `${stickyMessage.timeInterval}`, `${stickyMessage.messageCount}`, stickyMessage.channelId))] });
         }
 
         return msg.channel.send({ embeds: [helpEmbed(evokedCommand)] });
